fix(radspec): format time in delay script radspec description
